refactor(copyright): replace #[path] submodules with directory modules
Convert grammar and refiner from flat files using #[path = ...] to
idiomatic directory submodules. This improves discoverability and
aligns with Rust module conventions.

Grammar: grammar.rs + grammar_rules.rs + grammar_types.rs + grammar_test.rs
  → grammar/{mod,types,rules,tests}.rs

Refiner: refiner.rs + refiner_author.rs + refiner_utils.rs + refiner_test.rs
  → refiner/{mod,author,utils,tests}.rs

All external import paths remain unchanged. 473 tests pass.
